Make Mobius.RRD public
This exposes the `Mobius.RRD` module as a public API. The reason for
this change is when working with other software that uses Mobius I
noticed that I had to jump through extra steps to get the data from the
RRD. Also, this influenced some designed considerations in other parts
of the Mobius library.

Also, made the errors into exceptions as to establish the exception
pattern in Mobius. Since, RRD moving into a public API it seemed good to
move that direction now rather than later.