Refactor `Mobius.metric_name()` and metric name inconsistencies
This refactoring cleared up a few inconsistencies with what Mobius
expected as a metric name. In some spots it will a list of atoms, like
what is exposed in the `:telemetry` library, and other cases the string
version of the metric name. Also, a few places called the metric name
the event name, which in reality those things are different.

This refactoring exposing functions that use the atom list metric name
for areas that directly interact with `:telemetry`. Also the atom list
name is stored in the metrics table, as it is easier to write match
specs on a list of atoms. Everything else a user of Mobius will see
exposes the metric name as a string.

This also allowed the marshalling between two formats to be consolidated
to one place.