Support on demand reports
Update the remote reporter logic to allow for programmatically or
manually sending remote reports when requested.

This breaks the default behavior of sending reports every minute if no
interval is provided. To upgrade a user has to explicitly set the report
interval, otherwise Mobius will not automatically send reports.

This give more control over data usage for devices that use mobile
connects or systems that need to ingest data on demand rather than in real
time.