Refactor reporting functionality
Remove remote reporting support. This was causing a lot of extra
work for a feature that can completed better. I am removing this for now
as consumers can still use the building blocks to periodically report
to a remote server if that is their desire.

Added `Mobius.get_latest_events/1` and `Mobius.get_latest_metrics/1` to
enable consumers to call into Mobius to the latest data.